Working finally

+++ main.cpp
@@ -226,7 +226,7 @@ int main(int argc, char** args)
 	const int divisions = 20;
 	const int number_vertices = 3 * divisions * divisions;
 	const int number_indicies = 6 * (divisions-1) * (divisions - 1);
-	const int size = 100;
+	const float size = 100.0f;
 	GLfloat *ground_data = generateGround(-size, size, -size, size, divisions);
 	GLuint *indices = generateIndices(divisions);
